services/company-service/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
asyncpg
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
email-validator==2.3.0
python-multipart==0.0.24
psycopg2-binary==2.9.11
httpx==0.28.1
pgvector==0.4.2
cloudinary==1.44.1